Attemps to fix bug concerning displaying of unfinished tile improvements. Unfortunatly this doesn't work as expected as the graphics of the unfinished improvements aren't updaed as exspected.
Two new flags for the terrainDB to allow modders to customize the graphics for
goody huts, this modification allows the modder to define for each terrain a
different hut graphic, for example you could have glacier like style huts for
icy terrains or a dessert or oasis like ruin/hut for desserts and so on.

The new flags in the terrainDB are optional if both of these flags are missing
or just one of them a default image is used, like it is in the original CTP2.

The terrain.txt here is a modified version, I replaced some of the ruins
graphics with another graphic that is part of the tile file already.
This terrain.txt is meant as a base for disscussion, so far it just 
demonstraded that it does work.