Slic builtin update by Martin G-Aühmann:-b
..\ctp2_code\gs\slic\SlicBuiltin.cpp
The following builtins have been updated:

- player.capital: The player's capital as string or city_t.
- player.largestcity: The player's largest city as string or city_t.
- player.researching: The database index or name of the current player's 
  research project.
- player.militarystate: The player's military state as index or string.
- player.leaderperonality: The player's leader's personality type as
  string or personality database index.
- player.govttype: The player's government type as string or government
  database index.

..\ctp2_code\gs\slic\SlicSymbol.cpp
GetText function now returns FALSE to be able to check whether there 
according slic symbol has a valid string.

..\ctp2_code\gs\slic\sliccmd.cpp
Modifed the integer checker function so that they return FALSE if there is 
a valid string associated with the given slic symbol, so that the slic 
builtins above still work in string replacement as used.

..\ctp2_code\ai\diplomacy\sstateevent.cpp
Added two "!" so that the code for the AI to respect the city limit is used
in non-ACTIVISION_ORIGINAL builds instead the way around.