Fixed the EffectIncreaseScience feat not giving any science bonus. This bonus does not show in the empire manager yet. Cleaning up the empire manager will be a job by itself.
Fixed the EffectIncreaseHitPoints feat not giving any hit points bonus. This now works the same way as the wonder hit points effect. When the feat is built existing units get added the extra HP. When the feat is in effect new units start with the extra HP. An "undocumented feature" of this and the HP wonder effect is that when a unit loses health lower than it's maximum HP in units.txt, it can only heal back to it's original HP from units.txt. (Unless of course you build another HP wonder or achieve another HP feat). TODO: Make this bonus maximum hit points permanent until the wonder or feat effect is gone, also modify the interface or the green HP bar not to overlap the interface.